simple_rag_with_chroma_openai/requirements.txt
langchain-community
langchain-core
langchain-chroma
langchain-openai
openai-chat
chromadb
pypdf
flask

